+++ components/tvplanit/source/vpbaseds.pas
@@ -239,7 +239,7 @@ type
 
     procedure DeregisterAllWatchers;
     procedure DeregisterWatcher (Watcher : THandle);
-    function GetNextID(TableName: string): Int64; virtual; abstract;
+    function GetNextID(TableName: string): Integer; virtual; abstract;
     property Resources: TVpResources read FResources;
     procedure Load; virtual;
     procedure LoadEvents; virtual; abstract;
